add Makefile and README
A Makefile is now included in the project. This means that, to compile the
project, you don't need to do anything other than typing:

```
make
```

Any future additions, such as flags for `pdflatex`, can easily be added to the
Makefile. Super easy.

I've also added a README with instructions on building the project.